Add python strategy
