+++ src/transtructiver/mutation/rules/control_structure_substitution/control_structure_strategies/base_control_structure_strategy.py
@@ -26,6 +26,84 @@ class BaseControlStructureStrategy(ABC):
         apply(node, context, rule): Performs the transformation.
     """
 
+    def _insert_node(self, context, parent, point, type, text, index, rule):
+        """
+        Creates and inserts a Node, then returns the generated insertion record.
+
+        Args:
+            context: The execution context providing unique IDs.
+            parent: The parent Node receiving the new child.
+            point: The location where the new node will be inserted.
+            type (str): The node type for the Node and the record.
+            text (str): The text content for the Node and the record.
+            index (int): The position to insert the child in the parent's list.
+            records (list): The list where the record is stored.
+            rule: The rule object used to generate the record.
+
+        Returns:
+            dict: The record object generated by rule.record_insert.
+        """
+        new_node = Node(
+            start_point=(context.next_id(), -1),
+            end_point=point,
+            type=type,
+            text=text,
+        )
+        new_node.parent = parent
+        parent.children.insert(index, new_node)
+
+        record = rule.record_insert(
+            point=new_node.start_point,
+            insertion_point=point,
+            new_text=new_node.text,
+            new_type=type,
+        )
+
+        return record
+
+    def _insert_segments(self, context, parent, point, segments, index, rule):
+        """
+        Inserts a batch of nodes and returns the resulting records.
+
+        Args:
+            context: The execution context providing unique IDs.
+            parent: The parent Node receiving the new children.
+            point: The location where the new nodes will be inserted.
+            segments (list[tuple]): A list of (type, text) tuples to insert.
+            index (int): The starting position for insertions.
+            rule: The rule object used to generate records.
+
+        Returns:
+            list: A list of record objects generated during insertion.
+        """
+        records = []
+        for node_type, content in segments:
+            records.append(
+                self._insert_node(context, parent, point, node_type, content, index, rule)
+            )
+
+        return records
+
+    def _substitute(
+        self, node: Node, new_type: str, new_text: str, rule: MutationRule
+    ) -> MutationRecord:
+        """
+        Mutates node type/text and returns the substitution record.
+
+        Args:
+            node: The node to mutate.
+            new_type: The new type to assign to the node.
+            new_text: The new text to assign to the node.
+            rule: The mutation rule instance used for recording the substitution.
+
+        Returns:
+            The generated mutation record.
+        """
+        old_type = node.type
+        node.type = new_type
+        node.text = new_text
+        return rule.record_substitute(node, old_type)
+
     @abstractmethod
     def is_valid(self, node: Node) -> bool:
         """
